cgi/cgiutils: Silence a compiler warning
Why a char can't be a an array subscript I have no idea, but
casting it to an int at least makes gcc shut up.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2392 5f96b256-904b-4d8d-8c98-d829582c6739

+++ cgi/cgiutils.c
@@ -412,7 +412,7 @@ int read_cgi_config_file(char *filename) {
 		}
 
 	for(p = illegal_output_chars; p && *p; p++)
-		illegal_output_char_map[*p] = 1;
+		illegal_output_char_map[(int)*p] = 1;
 
 	/* free memory and close the file */
 	free(input);
